backend/requirements.txt
flask
flask-socketio
pandas
numpy
tensorflow
tf-keras
scikit-learn
joblib
transformers
sqlalchemy
web3
gunicorn
requests
torch
three
matplotlib
seaborn
latex
plotly
networkx
reportlab
hedera-sdk-py
python-dotenv